+++ src/common.rs
@@ -1,9 +1,10 @@
 //! Common functionality and types.
 
-use std::path::PathBuf;
+use std::io::ErrorKind;
+use std::path::{Path, PathBuf};
 
 use anyhow::{anyhow, Context, Result};
-use async_std::path::PathBuf as AsyncPathBuf;
+use async_std::fs;
 use async_std::task::spawn_blocking;
 
 use console::Emoji;
@@ -23,10 +24,11 @@ pub fn parse_public_url(val: &str) -> String {
 
 /// A utility function to recursively copy a directory.
 pub async fn copy_dir_recursive(from_dir: PathBuf, to_dir: PathBuf) -> Result<()> {
-    if !AsyncPathBuf::from(&from_dir).exists().await {
+    if !path_exists(&from_dir).await? {
         return Err(anyhow!("directory can not be copied as it does not exist {:?}", &from_dir));
     }
-    spawn_blocking(move || {
+
+    spawn_blocking(move || -> Result<()> {
         let opts = fs_extra::dir::CopyOptions {
             overwrite: true,
             content_only: true,
@@ -36,6 +38,32 @@ pub async fn copy_dir_recursive(from_dir: PathBuf, to_dir: PathBuf) -> Result<()
         Ok(())
     })
     .await
+    .context("error copying directory")
+}
+
+/// A utility function to recursively delete a directory.
+///
+/// Use this instead of fs::remove_dir_all(...) because of Windows compatibility issues, per
+/// advice of https://blog.qwaz.io/chat/issues-of-rusts-remove-dir-all-implementation-on-windows
+pub async fn remove_dir_all(from_dir: PathBuf) -> Result<()> {
+    if !path_exists(&from_dir).await? {
+        return Ok(());
+    }
+    spawn_blocking(move || {
+        ::remove_dir_all::remove_dir_all(from_dir.as_path()).context("error removing directory")?;
+        Ok(())
+    })
+    .await
+}
+
+/// Checks if path exists.
+pub async fn path_exists(path: impl AsRef<Path>) -> Result<bool> {
+    let exists = fs::metadata(path.as_ref())
+        .await
+        .map(|_| true)
+        .or_else(|error| if error.kind() == ErrorKind::NotFound { Ok(false) } else { Err(error) })
+        .with_context(|| format!("error checking for existance of path at {:?}", path.as_ref()))?;
+    Ok(exists)
 }
 
 /// Build system spinner.