+++ src/pipelines/rust_app.rs
@@ -1,20 +1,21 @@
 //! Rust application pipeline.
 
-use std::path::PathBuf;
+use std::iter::Iterator;
+use std::path::{Path, PathBuf};
 use std::sync::Arc;
+use std::{ffi::OsStr, str::FromStr};
 
-use anyhow::{anyhow, ensure, Context, Result};
+use anyhow::{anyhow, bail, Context, Result};
 use async_process::{Command, Stdio};
 use async_std::fs;
-use async_std::path::Path;
 use async_std::task::{spawn, JoinHandle};
 use futures::channel::mpsc::Sender;
 use indicatif::ProgressBar;
-use nipper::{Document, Selection};
+use nipper::Document;
 
-use super::TrunkLinkPipelineOutput;
+use super::{LinkAttrs, TrunkLinkPipelineOutput};
 use super::{ATTR_HREF, SNIPPETS_DIR};
-use crate::common::copy_dir_recursive;
+use crate::common::{copy_dir_recursive, path_exists};
 use crate::config::{CargoMetadata, RtcBuild};
 
 /// A Rust application pipeline.
@@ -27,24 +28,90 @@ pub struct RustApp {
     progress: ProgressBar,
     /// All metadata associated with the target Cargo project.
     manifest: CargoMetadata,
-    /// An optional channel to be used to communicate ignore paths to the watcher.
+    /// An optional channel to be used to communicate paths to ignore back to the watcher.
     ignore_chan: Option<Sender<PathBuf>>,
-    /// An optional binary name which will cause cargo & wasm-bindgen to process only the target binary.
+    /// An optional binary name which will cause cargo & wasm-bindgen to process only the target
+    /// binary.
     bin: Option<String>,
+    /// An optional optimization setting that enables wasm-opt. Can be nothing, `0` (default), `1`,
+    /// `2`, `3`, `4`, `s or `z`. Using `0` disables wasm-opt completely.
+    wasm_opt: WasmOptLevel,
+}
+
+/// Different optimization levels that can be configured with `wasm-opt`.
+#[derive(PartialEq, Eq)]
+enum WasmOptLevel {
+    /// Default optimization passes.
+    Default,
+    /// No optimization passes, skipping the wasp-opt step.
+    Off,
+    /// Run quick & useful optimizations. useful for iteration testing.
+    One,
+    /// Most optimizations, generally gets most performance.
+    Two,
+    /// Spend potentially a lot of time optimizing.
+    Three,
+    /// Also flatten the IR, which can take a lot more time and memory, but is useful on more nested
+    /// / complex / less-optimized input.
+    Four,
+    /// Default optimizations, focus on code size.
+    S,
+    /// Default optimizations, super-focusing on code size.
+    Z,
+}
+
+impl FromStr for WasmOptLevel {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s {
+            "" => Self::Default,
+            "0" => Self::Off,
+            "1" => Self::One,
+            "2" => Self::Two,
+            "3" => Self::Three,
+            "4" => Self::Four,
+            "s" | "S" => Self::S,
+            "z" | "Z" => Self::Z,
+            _ => bail!("unknown wasm-opt level `{}`", s),
+        })
+    }
+}
+
+impl AsRef<str> for WasmOptLevel {
+    fn as_ref(&self) -> &str {
+        match self {
+            Self::Default => "",
+            Self::Off => "0",
+            Self::One => "1",
+            Self::Two => "2",
+            Self::Three => "3",
+            Self::Four => "4",
+            Self::S => "s",
+            Self::Z => "z",
+        }
+    }
+}
+
+impl Default for WasmOptLevel {
+    fn default() -> Self {
+        // Current default is off until automatic download of wasm-opt is implemented.
+        Self::Off
+    }
 }
 
 impl RustApp {
     pub const TYPE_RUST_APP: &'static str = "rust";
 
     pub async fn new(
-        cfg: Arc<RtcBuild>, progress: ProgressBar, html_dir: Arc<PathBuf>, ignore_chan: Option<Sender<PathBuf>>, el: Selection<'_>, id: usize,
+        cfg: Arc<RtcBuild>, progress: ProgressBar, html_dir: Arc<PathBuf>, ignore_chan: Option<Sender<PathBuf>>, attrs: LinkAttrs, id: usize,
     ) -> Result<Self> {
         // Build the path to the target asset.
-        let manifest_href = el
-            .attr(ATTR_HREF)
-            .map(|tendril| {
+        let manifest_href = attrs
+            .get(ATTR_HREF)
+            .map(|attr| {
                 let mut path = PathBuf::new();
-                path.extend(tendril.as_ref().split('/'));
+                path.extend(attr.split('/'));
                 if !path.is_absolute() {
                     path = html_dir.join(path);
                 }
@@ -54,7 +121,8 @@ impl RustApp {
                 path
             })
             .unwrap_or_else(|| html_dir.join("Cargo.toml"));
-        let bin = el.attr("data-bin").map(|val| val.to_string());
+        let bin = attrs.get("data-bin").map(|val| val.to_string());
+        let wasm_opt = attrs.get("data-wasm-opt").map(|val| val.parse()).transpose()?.unwrap_or_default();
         let manifest = CargoMetadata::new(&manifest_href).await?;
         let id = Some(id);
 
@@ -65,6 +133,7 @@ impl RustApp {
             manifest,
             ignore_chan,
             bin,
+            wasm_opt,
         })
     }
 
@@ -80,6 +149,7 @@ impl RustApp {
             manifest,
             ignore_chan,
             bin: None,
+            wasm_opt: WasmOptLevel::default(),
         })
     }
 
@@ -90,19 +160,23 @@ impl RustApp {
 
     async fn build(mut self) -> Result<TrunkLinkPipelineOutput> {
         let (wasm, hashed_name) = self.cargo_build().await?;
-        let output = self.wasm_bindgen_build(wasm, hashed_name).await?;
+        let output = self.wasm_bindgen_build(wasm.as_ref(), &hashed_name).await?;
+        self.wasm_opt_build(&output.wasm_output).await?;
         Ok(TrunkLinkPipelineOutput::RustApp(output))
     }
 
     async fn cargo_build(&mut self) -> Result<(PathBuf, String)> {
         self.progress.set_message(&format!("building {}", &self.manifest.package.name));
-        if let Some(chan) = &mut self.ignore_chan {
-            let _ = chan.try_send(self.manifest.metadata.target_directory.clone());
-        }
 
         // Spawn the cargo build process.
+        let color_mode = if atty::is(atty::Stream::Stdout) {
+            "--color=always"
+        } else {
+            "--color=never"
+        };
         let mut args = vec![
             "build",
+            color_mode,
             "--target=wasm32-unknown-unknown",
             "--manifest-path",
             &self.manifest.manifest_path,
@@ -123,11 +197,19 @@ impl RustApp {
             .output()
             .await
             .context("error during cargo build execution")?;
-        ensure!(
-            build_output.status.success(),
-            "bad status returned from cargo build: {}",
-            String::from_utf8_lossy(&build_output.stderr)
-        );
+
+        // Send cargo's target dir over to the watcher to be ignored. We must do this before
+        // checking for errors, otherwise the dir will never be ignored. If we attempt to do
+        // this pre-build, the canonicalization will fail and will not be ignored.
+        if let Some(chan) = &mut self.ignore_chan {
+            let _ = chan.try_send(self.manifest.metadata.target_directory.clone());
+        }
+
+        // Now propagate any errors which came from the cargo build.
+        if !build_output.status.success() {
+            self.progress.println(String::from_utf8_lossy(&build_output.stderr));
+            bail!("bad status returned from cargo build");
+        }
 
         // Perform a final cargo invocation on success to get artifact names.
         self.progress.set_message("fetching artifacts");
@@ -141,11 +223,10 @@ impl RustApp {
             .output()
             .await
             .context("error getting cargo build artifacts info")?;
-        ensure!(
-            artifacts_out.status.success(),
-            "bad status returned from cargo artifacts request: {}",
-            String::from_utf8_lossy(&build_output.stderr)
-        );
+        if !artifacts_out.status.success() {
+            self.progress.println(String::from_utf8_lossy(&artifacts_out.stderr));
+            bail!("bad status returned from cargo artifacts request");
+        }
 
         // Stream over cargo messages to find the artifacts we are interested in.
         let reader = std::io::BufReader::new(artifacts_out.stdout.as_slice());
@@ -156,24 +237,24 @@ impl RustApp {
                 cargo_metadata::Message::BuildFinished(finished) if !finished.success => Err(anyhow!("error while fetching cargo artifact info")),
                 _ => acc,
             })?
-            .ok_or_else(|| anyhow!("cargo artifacts not found for target crate"))?;
+            .context("cargo artifacts not found for target crate")?;
 
         // Get a handle to the WASM output file.
         let wasm = artifact
             .filenames
             .into_iter()
             .find(|path| path.extension().map(|ext| ext == "wasm").unwrap_or(false))
-            .ok_or_else(|| anyhow!("could not find WASM output after cargo build"))?;
+            .context("could not find WASM output after cargo build")?;
 
         // Hash the built wasm app, then use that as the out-name param.
         self.progress.set_message("processing WASM");
-        let wasm_bytes = async_std::fs::read(&wasm).await.context("error reading wasm file for hash generation")?;
+        let wasm_bytes = fs::read(&wasm).await.context("error reading wasm file for hash generation")?;
         let hashed_name = format!("index-{:x}", seahash::hash(&wasm_bytes));
         Ok((wasm, hashed_name))
     }
 
-    async fn wasm_bindgen_build(&self, wasm: PathBuf, hashed_name: String) -> Result<RustAppOutput> {
-        self.progress.set_message("calling wasm-bindgen");
+    async fn wasm_bindgen_build(&self, wasm: &Path, hashed_name: &str) -> Result<RustAppOutput> {
+        self.progress.set_message("preparing for build");
 
         // Ensure our output dir is in place.
         let mode_segment = if self.cfg.release { "release" } else { "debug" };
@@ -189,40 +270,31 @@ impl RustApp {
         let args = vec!["--target=web", &arg_out_path, &arg_out_name, "--no-typescript", &target_wasm];
 
         // Invoke wasm-bindgen.
-        let build_output = Command::new("wasm-bindgen")
-            .args(args.as_slice())
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .spawn()
-            .context("error spawning wasm-bindgen call")?
-            .output()
-            .await
-            .context("error during wasm-bindgen call")?;
-        ensure!(
-            build_output.status.success(),
-            "wasm-bindgen call returned a bad status {}",
-            String::from_utf8_lossy(&build_output.stderr),
-        );
+        self.progress.set_message("calling wasm-bindgen");
+        run_command("wasm-bindgen", &args, self.progress.clone()).await?;
+
+        self.progress.set_message("copying generated artifacts");
 
         // Copy the generated WASM & JS loader to the dist dir.
-        self.progress.set_message("copying generated artifacts");
         let hashed_js_name = format!("{}.js", &hashed_name);
         let hashed_wasm_name = format!("{}_bg.wasm", &hashed_name);
         let js_loader_path = bindgen_out.join(&hashed_js_name);
-        let js_loader_path_dist = self.cfg.dist.join(&hashed_js_name);
+        let js_loader_path_dist = self.cfg.staging_dist.join(&hashed_js_name);
         let wasm_path = bindgen_out.join(&hashed_wasm_name);
-        let wasm_path_dist = self.cfg.dist.join(&hashed_wasm_name);
+        let wasm_path_dist = self.cfg.staging_dist.join(&hashed_wasm_name);
         fs::copy(js_loader_path, js_loader_path_dist)
             .await
-            .context("error copying JS loader file to dist dir")?;
-        fs::copy(wasm_path, wasm_path_dist).await.context("error copying wasm file to dist dir")?;
+            .context("error copying JS loader file to stage dir")?;
+        fs::copy(wasm_path, wasm_path_dist)
+            .await
+            .context("error copying wasm file to stage dir")?;
 
         // Check for any snippets, and copy them over.
         let snippets_dir = bindgen_out.join(SNIPPETS_DIR);
-        if Path::new(&snippets_dir).exists().await {
-            copy_dir_recursive(bindgen_out.join(SNIPPETS_DIR), self.cfg.dist.join(SNIPPETS_DIR))
+        if path_exists(&snippets_dir).await? {
+            copy_dir_recursive(bindgen_out.join(SNIPPETS_DIR), self.cfg.staging_dist.join(SNIPPETS_DIR))
                 .await
-                .context("error copying snippets dir")?;
+                .context("error copying snippets dir to stage dir")?;
         }
 
         Ok(RustAppOutput {
@@ -231,6 +303,38 @@ impl RustApp {
             js_output: hashed_js_name,
             wasm_output: hashed_wasm_name,
         })
+    }
+
+    async fn wasm_opt_build(&self, hashed_name: &str) -> Result<()> {
+        // If opt level is off, we skip calling wasm-opt as it wouldn't have any effect.
+        if self.wasm_opt == WasmOptLevel::Off {
+            return Ok(());
+        }
+
+        self.progress.set_message("calling wasm-opt");
+
+        // Ensure our output dir is in place.
+        let mode_segment = if self.cfg.release { "release" } else { "debug" };
+        let output = self.manifest.metadata.target_directory.join("wasm-opt").join(mode_segment);
+        fs::create_dir_all(&output).await.context("error creating wasm-opt output dir")?;
+
+        // Build up args for calling wasm-opt.
+        let output = output.join(hashed_name);
+        let arg_output = format!("--output={}", output.display());
+        let arg_opt_level = format!("-O{}", self.wasm_opt.as_ref());
+        let target_wasm = self.cfg.staging_dist.join(hashed_name).to_string_lossy().to_string();
+        let args = vec![&arg_output, &arg_opt_level, &target_wasm];
+
+        // Invoke wasm-opt.
+        run_command("wasm-opt", &args, self.progress.clone()).await?;
+
+        // Copy the generated WASM file to the dist dir.
+        self.progress.set_message("copying generated artifacts");
+        fs::copy(output, self.cfg.staging_dist.join(hashed_name))
+            .await
+            .context("error copying wasm file to dist dir")?;
+
+        Ok(())
     }
 }
 
@@ -260,4 +364,23 @@ impl RustAppOutput {
         }
         Ok(())
     }
+}
+
+/// Run a global command with the given arguments and make sure it completes successfully. If it
+/// fails an error is returned.
+async fn run_command(name: &str, args: &[impl AsRef<OsStr>], progress: ProgressBar) -> Result<()> {
+    let output = Command::new(name)
+        .args(args)
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .spawn()
+        .with_context(|| format!("error spawning {} call", name))?
+        .output()
+        .await
+        .with_context(|| format!("error during {} call", name))?;
+    if !output.status.success() {
+        progress.println(String::from_utf8_lossy(&output.stderr));
+        bail!("{} call returned a bad status", name);
+    }
+    Ok(())
 }