+++ src/watch.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{Context, Result};
 use async_std::task::{spawn_blocking, JoinHandle};
 use crossbeam_channel::Sender as CrossbeamSender;
 use futures::channel::mpsc::{channel, Receiver, Sender};
@@ -19,7 +19,7 @@ pub struct WatchSystem {
     /// The build system.
     build: BuildSystem,
     /// The current vector of paths to be ignored.
-    ignores: Vec<PathBuf>,
+    ignored_paths: Vec<PathBuf>,
     /// A channel of FS watch events.
     watch_rx: Receiver<DebouncedEvent>,
     /// A channel of new paths to ignore from the build system.
@@ -35,23 +35,15 @@ impl WatchSystem {
         let (watch_tx, watch_rx) = channel(1);
         let (build_tx, build_rx) = channel(1);
 
-        // Process ignore list.
-        let mut ignores = cfg.ignore.iter().try_fold(vec![], |mut acc, path| -> Result<Vec<PathBuf>> {
-            let abs_path = path.canonicalize().map_err(|err| anyhow!("invalid path provided: {}", err))?;
-            acc.push(abs_path);
-            Ok(acc)
-        })?;
-        ignores.append(&mut vec![cfg.build.dist.clone()]);
-
         // Build the watcher.
-        let _watcher = build_watcher(watch_tx)?;
+        let _watcher = build_watcher(watch_tx, cfg.paths.clone())?;
 
         // Build dependencies.
         let build = BuildSystem::new(cfg.build.clone(), progress.clone(), Some(build_tx), build_event_tx).await?;
         Ok(Self {
             progress,
             build,
-            ignores,
+            ignored_paths: cfg.ignored_paths.clone(),
             watch_rx,
             build_rx,
             _watcher,
@@ -81,37 +73,61 @@ impl WatchSystem {
     }
 
     async fn handle_watch_event(&mut self, event: DebouncedEvent) {
-        let ev_path = match event {
+        let mut ev_path = match event {
             DebouncedEvent::Create(path) | DebouncedEvent::Write(path) | DebouncedEvent::Remove(path) | DebouncedEvent::Rename(_, path) => path,
             _ => return,
         };
-        for path in ev_path.ancestors() {
-            if self.ignores.iter().map(|p| p.as_path()).any(|p| p == path) {
-                return; // Don't emit a notification if ignored.
-            }
+
+        ev_path = match ev_path.canonicalize() {
+            Ok(path) => path,
+            // Ignore errors here, as this would only take place for a resource which has
+            // been removed, which will happen for each of our dist/.stage entries.
+            Err(_) => return,
+        };
+
+        if ev_path
+            .ancestors()
+            .any(|path| self.ignored_paths.iter().any(|ignored_path| ignored_path == path))
+        {
+            return; // Don't emit a notification if path is ignored.
         }
+
         if let Err(err) = self.build.build().await {
-            self.progress.println(format!("{}", err));
+            self.progress.println(format!("{:?}", err));
         }
     }
 
-    fn update_ignore_list(&mut self, path: PathBuf) {
-        if !self.ignores.contains(&path) {
-            self.ignores.push(path);
+    fn update_ignore_list(&mut self, arg_path: PathBuf) {
+        let path = match arg_path.canonicalize() {
+            Ok(canon_path) => canon_path,
+            Err(_) => arg_path,
+        };
+
+        if !self.ignored_paths.contains(&path) {
+            self.ignored_paths.push(path);
         }
     }
 }
 
-fn build_watcher(mut watch_tx: Sender<DebouncedEvent>) -> Result<(JoinHandle<()>, RecommendedWatcher)> {
+fn build_watcher(mut watch_tx: Sender<DebouncedEvent>, paths: Vec<PathBuf>) -> Result<(JoinHandle<()>, RecommendedWatcher)> {
     let (tx, rx) = std::sync::mpsc::channel();
     let mut watcher = watcher(tx, std::time::Duration::from_secs(1)).context("failed to build file system watcher")?;
-    watcher
-        .watch(".", RecursiveMode::Recursive)
-        .context("failed to watch CWD for file system changes")?;
+
+    // Create a recursive watcher on each of the given paths.
+    // NOTE WELL: it is expected that all given paths are canonical. The Trunk config
+    // system currently ensures that this is true for all data coming from the
+    // RtcBuild/RtcWatch/RtcServe/&c runtime config objects.
+    for path in paths {
+        watcher
+            .watch(&path, RecursiveMode::Recursive)
+            .context(format!("failed to watch {:?} for file system changes", path))?;
+    }
+
     let handle = spawn_blocking(move || loop {
         if let Ok(event) = rx.recv() {
             let _ = watch_tx.try_send(event);
         }
     });
+
     Ok((handle, watcher))
 }